validate: diversify valid examples for printing
otherwise we only see the longest examples. this is esp. helpful in bitod since there are long texts for directions.
